Update email files and polish up for deployment
