+++ app/Library/Lesson/Enroll.php
@@ -5,8 +5,8 @@ namespace App\Library\Lesson;
 use App\Lesson;
 use App\Library\NewsLetter\NewsLetter;
 use App\Library\StripeCharge;
-use App\Mail\ClassFull;
-use App\Mail\SignUp;
+use App\Mail\Groups\ClassFull;
+use App\Mail\Privates\SignUp;
 use App\Swimmer;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Log;
