+++ app/Mail/Groups/ClassFull.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace App\Mail;
+namespace App\Mail\Groups;
 
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
 use Illuminate\Queue\SerializesModels;
+use Illuminate\Contracts\Queue\ShouldQueue;
 use App\Lesson;
 
-class SignUp extends Mailable
+class ClassFull extends Mailable
 {
     use Queueable, SerializesModels;
 
@@ -18,7 +19,7 @@ class SignUp extends Mailable
 
 
     /**
-     * SignUp constructor.
+     * ClassFull constructor.
      * @param Lesson $lesson
      */
     public function __construct(Lesson $lesson)
@@ -33,8 +34,7 @@ class SignUp extends Mailable
      */
     public function build()
     {
-        return $this->subject($this->lesson->group->type)
-                    ->markdown('email.lessonSignUp')
-                    ->with(['lesson', $this->lesson]);
+        return $this->subject($this->lesson->group->type.' lesson is full.')
+                    ->markdown('email.classFull');
     }
 }
