+++ app/Mail/NewsLetter/RegistrationOpen.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Mail;
+namespace App\Mail\NewsLetter;
 
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
