--- app/Mail/RegistrationOpeningSoon.php
@@ -1,39 +1,0 @@
-<?php
-
-namespace App\Mail;
-
-use Illuminate\Bus\Queueable;
-use Illuminate\Mail\Mailable;
-use Illuminate\Queue\SerializesModels;
-use Illuminate\Contracts\Queue\ShouldQueue;
-
-class RegistrationOpeningSoon extends Mailable
-{
-    use Queueable, SerializesModels;
-
-    /**
-     * @var
-     */
-    public $emailAddress;
-
-    /**
-     * Create a new message instance.
-     * @param $emailAddress
-     * @return void
-     */
-    public function __construct($emailAddress)
-    {
-        $this->emailAddress = $emailAddress;
-    }
-
-    /**
-     * Build the message.
-     *
-     * @return $this
-     */
-    public function build()
-    {
-        return $this->markdown('email.registrationOpeningSoon')
-                    ->with(['emailAddress', $this->emailAddress]);
-    }
-}
