+++ app/Mail/SwimTeam/TryoutReminder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Mail;
+namespace App\Mail\SwimTeam;
 
 use App\Tryout;
 use Illuminate\Bus\Queueable;
