+++ phpunit.xml
@@ -22,6 +22,9 @@
             <directory suffix=".php">./app</directory>
         </whitelist>
     </filter>
+    <listeners>
+        <listener class="NunoMaduro\Collision\Adapters\Phpunit\Listener" />
+    </listeners>
     <php>
         <env name="APP_ENV" value="testing"/>
         <env name="TELESCOPE_ENABLED" value="false"/>
