+++ resources/views/pages/about.blade.php
@@ -36,7 +36,7 @@ About The Swim School
                 <hr class="uk-divider-icon uk-scrollspy-inview uk-animation-slide-top-medium" uk-scrollspy-class="">
                 <div class="uk-margin uk-dropcap">
                     <p>
-                        Originally from Pennsylvania, Hilary started swimming at the age of five and continued competitively through college in Ohio where she earned a B.A. in Psychology. After receiving her degree in 2003, she returned to PA and worked part-time as a lifeguard, <a title="Bradenton Swim Instruction" href="{{ route('privates.index') }}">swim lesson instructor</a>, and swim team coach in addition to having a full-time position in the human services field assisting children with special needs. With a love for the beach and a craving for a warmer climate, Hilary relocated to Florida in 2006 and worked as an Aquatics Director at a YMCA for over eight years where she developed a year round competitive swim team program, administered a progressive swim lesson program and conducted water fitness classes at two aquatics facilities. She is a certified AEA Aquatic Fitness Professional, YMCA Swim Lesson Instructor, YMCA Lifeguard, and ASHI CPR-PR/AED & First Aid Instructor.
+                        Originally from Pennsylvania, Hilary started swimming at the age of five and continued competitively through college in Ohio where she earned a B.A. in Psychology. After receiving her degree in 2003, she returned to PA and worked part-time as a lifeguard, <a title="Bradenton Swim Instruction" href="{{ route('private_lesson.index') }}">swim lesson instructor</a>, and swim team coach in addition to having a full-time position in the human services field assisting children with special needs. With a love for the beach and a craving for a warmer climate, Hilary relocated to Florida in 2006 and worked as an Aquatics Director at a YMCA for over eight years where she developed a year round competitive swim team program, administered a progressive swim lesson program and conducted water fitness classes at two aquatics facilities. She is a certified AEA Aquatic Fitness Professional, YMCA Swim Lesson Instructor, YMCA Lifeguard, and ASHI CPR-PR/AED & First Aid Instructor.
                     </p>
                 </div>
             </div>
